requirements.txt
Flask==2.2.2
huggingface-hub==0.10.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
scikit-learn
scipy==1.9.3
seaborn==0.12.1
tensorboard==2.10.1
torch==1.13
torchaudio==0.13
torchvision==0.14
tqdm==4.64.1
transformers==4.24.0

